removida a sombra pirosa, fonte correta e ficheiros organizados
