--- smartbreak/src/routes/routes.js
@@ -1,11 +1,0 @@
-// import { NavigationContainer } from "@react-navigation/native";
-
-// import { Tabbar } from './tabbar';
-
-// export function Routes() {
-//     return (
-//         <NavigationContainer>
-//            <Tabbar /> 
-//         </NavigationContainer>
-//     )
-// }
